added indexing of attachments
Attachments can be selected as to-be-indexed post types on the Solr
options page. If selected, they are parsed using Solr's extract handler
and indexed.
Some attachments lead to an uncaught exception and freeze the indexing
process. I have no solution yet, except putting these attachments on the
exclude items list.